Create nfa.cpp
There are two parts to this program, the first which I have working. The first portion should take a text file which has all the states of a machine. It records the final states, the beginning state, the number of total states, the number of letters in the alphabet used including what they are, and then all of the transitions from states to states. That works completely fine. The problem i am having is making a function that can test strings of input to see if they are accepted by the machine. I wanted to make a recursive function to do this but that means a lot of parameters have to be setup for the function. This would include the 3D array that stores all of the transitions, and i don't know how to pass a 3D array in a function because i won't know how big the array will be until this point in the program.  The size of the array depends on the number of states and number of letters in the alphabet.